+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.5.3]
 ### Changed
 - Throw a more readable exception if `$ref` or `allOf` are found during conversion
 - Assume something is object type during conversion if it has `properties`
@@ -127,8 +129,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2021-05-31
 - Initial release
 
-[Unreleased]: https://github.com/michaelmior/jsonoid-discovery/compare/v0.5.2...HEAD
-[0.5.1]: https://github.com/michaelmior/jsonoid-discovery/compare/v0.5.1...v0.5.2
+[Unreleased]: https://github.com/michaelmior/jsonoid-discovery/compare/v0.5.3...HEAD
+[0.5.3]: https://github.com/michaelmior/jsonoid-discovery/compare/v0.5.2...v0.5.3
+[0.5.2]: https://github.com/michaelmior/jsonoid-discovery/compare/v0.5.1...v0.5.2
 [0.5.1]: https://github.com/michaelmior/jsonoid-discovery/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/michaelmior/jsonoid-discovery/compare/v0.4.2...v0.5.0
 [0.4.2]: https://github.com/michaelmior/jsonoid-discovery/compare/v0.4.1...v0.4.2